Correctly kill subprocesses; allow with-blocks for connections
This allows ServerProcess, CryptolConnection, and SAWConnection to be used as
resources in with-blocks.

Further, we ensure that even on tricksy macOS, the subprocesses spawned are
appropriately killed on the end of a with-block or when the object is deleted.
This is done by opening the new process as a separate group, and killing the
entire group (with all its subprocesses), to get around what seems to be a
subtle difference in how the shell=True functionality interacts with
subprocesses on macOS vs Linux.

This has not yet been tested on Windows!