Allow Python to handle select Argo exceptions via process_error
If a subclass of Command or Query overrides the process_error method, which
takes (code : int, data : Any, msg : str), so that it returns other than None
sometimes, the non-None values are returned non-exceptionally instead of a
thrown ArgoException. This allows us, for example, to recover from and log a
verification exception and continue with the script.