Semi-persistent server in Python for auto-refreshing "live view"
This implements a "live view" functionality which stands up a simple server and
allows multiple separate processes to tell it what to serve at any given path,
by manipulating a single shared temporary directory. The server is ephemeral,
because the live-views provide a heartbeat to tell it that someone is still
looking: when that heartbeat disappears for a certain threshold of time, it will
shut itself down to prevent zombie processes. But, when another request comes
in, it will start up again.