Remove buggy timeout functionality
The server will just run forever, for now, since we have no way to reliably
cancel it without leaking broken file descriptors and setting up weird errors
for the future. So, instead, we'll just let it run forever or be manually
killed. In the future, we should use a different web server that can gracefully
shutdown after a certain timeout with no new requests.