Refactor Options type, avoiding validation and adding --new option
Since the Options type is now incapable of representing an invalid set of
options (previously --public --stdio was invalid but representable), all
validation occurs inside optparse-applicative.

Furthermore, there is now a specific option --new that specifies a new server
should be launched, rather than attempting to use an existing server. The
default will be to maintain a persistent server.