Take --session parameter to allow creating/attaching to named sessions
The --session command line parameter allows the user to specify a global name
for the session created by an invocation of the saw server. If a server is
already running with this session name, the port number of that server is
printed and the process exits. Otherwise, the port number selected (either
dynamically or not) is printed, and the server continues executing.

This commit also does away with the IPv4/IPv6 distinction, permitting
connections over both IPv4 and IPv6 for any invocation of the server. The
--public flag still is used to determine whether the server is bound to the
address `::` (publicly reachable) or `::1` (only reachable locally).

The global information about sessions is stored in ~/.argo/sessions/, or its
idiomatic equivalent on other systems (on Windows, this should be something like
C:\Users\<user>\AppData\Roaming\argo\sessions).

At present, there is not a cleanup mechanism for the files left behind by
sessions. Each unique session name will create a few zero-size lock files and a
single `.port` file with a few bytes in it, so this leak is likely to be quite
slow, but it is still worth considering how to prevent it. Perhaps on startup,
the server should clean up all the files for any inactive sessions prior to
doing work itself?