Capture std{out|err} of methods and silence them
Previously, methods were themselves able to use stdout/stderr, which led to
broken pipe errors when a server process was started by a process (e.g. the
Python interpreter) which then quit whilst leaving the server persistent. To
prevent this issue, the stdout/stderr of any method invoked during the lifetime
of the server is captured, redirected to a temp file during the execution of
that method, and then loaded back. Currently we don't use this information, but
in the future we could send these pieces of text back to the client of the
server.