Attempt to output the stdout/stderr of methods, matching previous behavior
When the originally attached stdout/stderr exist, we would like to be able to
see what's been printed from within the method. Now, the stdout/stderr are still
captured, but after the method is finished being invoked, we *attempt* to
replicate the captured strings to the standard handles; if that fails,
though (i.e. the pipe is broken), we swallow the error, since this is an
expected scenario sometimes.