Adapt visual dashboard (using myxine) to use new View interface
This relies on a version of myxine which does not use the diffHTML package,
because diffing SVGs doesn't work there. A future release may include a
command-line flag to toggle using a bare assignment to document.body.innerHTML,
but until then, this works only with the saw-api-demo branch of myxine.