Detangled dashboard (#75)
* Use myxine instead of native python server for showing dashboard

* Add loading animation

* Remove redundant svg-wrapping in HTML result

* Bump stack LTS

* Replace connect commands in examples with cabal-run

* Capture std{out|err} of methods and silence them

Previously, methods were themselves able to use stdout/stderr, which led to
broken pipe errors when a server process was started by a process (e.g. the
Python interpreter) which then quit whilst leaving the server persistent. To
prevent this issue, the stdout/stderr of any method invoked during the lifetime
of the server is captured, redirected to a temp file during the execution of
that method, and then loaded back. Currently we don't use this information, but
in the future we could send these pieces of text back to the client of the
server.

* Attempt to output the stdout/stderr of methods, matching previous behavior

When the originally attached stdout/stderr exist, we would like to be able to
see what's been printed from within the method. Now, the stdout/stderr are still
captured, but after the method is finished being invoked, we *attempt* to
replicate the captured strings to the standard handles; if that fails,
though (i.e. the pipe is broken), we swallow the error, since this is an
expected scenario sometimes.

* Adapt visual dashboard (using myxine) to use new View interface

This relies on a version of myxine which does not use the diffHTML package,
because diffing SVGs doesn't work there. A future release may include a
command-line flag to toggle using a bare assignment to document.body.innerHTML,
but until then, this works only with the saw-api-demo branch of myxine.