Even when an exception is thrown, print stdout/err
Previously, when an error occurred, the error was indeed returned to the
client, but the stdout/err was swallowed since the corresponding print
statements were never executed for the captured output. This now is
fixed by catching and rethrowing exceptions at the appropriate place,
though first printing the captured output to the appropriate handles.