Modify Argo protocol and Python SAW client to track stdout/err
The Argo protocol now transmits the full captured stdout/err of the
method invoked back to the client, which may do with it what they
please. In the case of the Python client to the SAW server, this
information is exposed in the exceptions thrown, and in particular, to
the View implementations that can track and display this information as
if it was coming natively from the Python process.