Redo state management
States are now represented by opaque tokens (UUIDs) that are mapped
behind the scenes to a cache of application states as a mix of actual
reified saved states and recipes.

As part of getting all this to work, the submodules that are in common
with saw-script were set to just use those in saw-script.