Rework cache behavior
The prior version of this code assumed a bijective mapping between
valid state IDs and valid recipes. But this is not the case - one
recipe realised in separate states can give rise to separate state
IDs.

This version fixes that problem. In the process, the code became much
better-documented as well.