feat: bound memory usage by server, related testing
This PR modifies Argo servers to guarantee a bound on memory usage.
This is accomplished by:
  1. keeping only the initial server state cached indefinitely,
  2. keeping alive a (customizable) maximum number of recently used states
     while evicting others from the cache when it fills (along with
     any no longer in-use file contents loaded from disk), and
  3. allows states to be explicitly pinned in the cache, which will
     keep them in memory until they are explicitly unpinned.

The main cost of this change is that arbitrary states no longer
in cache cannot be "replayed" or "resurrected" automatically when
a request mentioning them is received once they are evicted from the cache
(such a state would either have to be manually pinned or rebuilt
by the client through standard interactions).