Python: Avoid typed-ast dependency
The `typed-ast` package is being end-of-life'd soon. See python/typed_ast#179.
Moreover, it is causing issues when building with Python 3.11+. See
GaloisInc/cryptol#1491.  Thankfully, `typed-ast` is no longer required if you
use sufficiently new versions of `mypy` and Python. This patch raises the lower
version bounds to make it possible to shed the `typed-ast` dependency entirely.