Add `customMain` argument for `--version`-like global options
`argo`'s allows users to customize the CLI option-parsing logic via the
`customMain` function. The primary means to do so is by adding options for each
subcommand (e.g., `stdio`). However, it is also convenient to allow users to
specify global options for things that look like a `--version` option—i.e.,
that print some information and exit.

This patch enables this functionality by adding an extra `(forall a. Parser (a
-> a))` argument to `customMain`, which can be used to configure global options
that resemble `--version`. I have also updated the `file-echo-api` example
applications in this repo to support a `--version` option, which provides a
demonstration of how this functionality can be used.

Fixes #218.